aya: remove RwLocks and Ref wrappers on bpf maps.
This improves performance by removing the runtime checking of borrowing
rules. In order to retain a similar level of flexibility as the previous
implementation permitted, `take_map` and `return_map` are added as
methods to Bpf, allowing callers to move ownership of a Map.

This also removes the need for the parking_lot dependency.