Add support for LSM programs attached to cgroups
Kernel 6.0 provides a new attachment type - `BPF_LSM_CGROUP`. When using
it, a program attaches to LSM hooks, but only within a cgroup.

`BPF_LSM_CGROUP` requires programs to we present in the `lsm_cgroup`
section, therefore we provide a new `cgroup_lsm` macro for defining such
programs in aya-bpf.

We also provide a new `CgroupLsm` structure in userspace aya, which uses
the new attachment type and stores the information about cgroup in
links.

Fixes: #423
Signed-off-by: Michal Rostecki <vadorovsky@gmail.com>