workflows: Run clippy for user space and eBPF crates separately
Clippy triggers builds for the checked crates. Without any additional
arguments, it does so with the default toolchain for the default target.
The default target for x86_64 hosts is x86_64-unknown-linux-gnu. Also,
lack of additional arguments means using the std library.

To avoid issues coming from compiling eBPF crates for x86_64 target and
with std (like [0]), this change ensures that they are checked with
`--target bpfel-unknown-none` and `-Zbuild-std=core` options.

[0] https://github.com/aya-rs/aya/actions/runs/4669677849/jobs/8268452592