aya-bpf: Add BTF maps (as a feature)
The new `btf-maps` feature enables usage of BTF-defined BPF map
definitions[0] instead of legacy ones.

This is still work in progress and it needs modification of debug info
with bpf-linker, which is being worked on in aya-rs/bpf-linker#35. This
draft PR is a reference for the bpf-linker work for now.

[0] https://lwn.net/Articles/790177/