xtask: ensure libbpf submodule is initialized
For xtask tasks which need libbpf, make sure it's initialized automatically.
This eases the user-experience so that users don't need to think about the submodule,
while retaining all the benefits of using a submodule vs forcing the user to manually
check out libbpf and stick it in some pre-defined place.