aya: Use AsFd when attaching fds to programs
This is a breaking change but adds another level of safety to ensure
the file descriptors we receive are valid. Additionally, this allows
aya to internally easily duplicate this file descriptors using std
library methods instead of manually calling `dup` which doesn't
duplicate with the CLOSE_ON_EXEC flag that is standard pratice to
avoid leaking file descriptors when exec'ing.