integration-test, xtask: Move xtask run docs into README
This moves the docs for downloading kernels into the integration-test
README since it allows us to be more verbose.

Instructions have been added for the use of Fedora kernels.

The instructions for Debian kernels were amended to not download all the
kernels, and to encourage targetted download of the one you care about.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>