chore(aya-ebpf)!: Deprecate aya-ebpf-cty
I don't recall why we used aya-ebpf-cty (a fork of the cty crate).
I assume it's because the ffi types in libcore weren't stablized when
the project first started.

Either way, I've replaced aya-ebpf-cty with core::ffi types and
the tests appear to pass still, which is reassuring.

This is a breaking change since the re-export of aya-ebpf-cty as
aya-ebpf::cty has also been removed. Migration is as simple as
replacing import with core::ffi.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>