chore(aya): Fix clippy unused cfg_attr
clippy complains that cfg_attr is applied to the macro invocation and
therefore will not be expanded. Attempted to fix this (and clean up
macro usage) by adding a meta matcher to these macros.

However, evaluating the expanded output shows that cfg_attr gets
ignored?

```rust

 #[cfg(any(feature = "async_tokio", feature = "async_std"))]
    impl<'a> TryFrom<&'a mut Map> for AsyncPerfEventArray<&'a mut MapData> {
        type Error = MapError;
        fn try_from(map: &'a mut Map) -> Result<Self, Self::Error> {
            match map {
                Map::PerfEventArray(map_data) => Self::new(map_data),
                map =>
                    Err(MapError::InvalidMapType { map_type: map.map_type() }),
            }
        }
    }
```

This supresses the clippy error for now, and the cfg_attr is only for
docs so it's not a big issue, but it would be nice to get this fixed
someday.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>